Migration: Interface refactoring for mypy fixes
- A recent mypy update seems to have caused all sort of new issues- this
fixes them
